Reorganizacion para implementar repositorios e implementación de los modulos sortable, nestable sobre el repo en lugar de sobre el modelo.

+++ Journal2API/Models/Comment.cs
@@ -8,13 +8,13 @@ namespace Journal2API.Models
 {
     public class Comment : HasTimestamp, IItem
     {
-        public int Id { get; set; }
+        public ulong Id { get; set; }
         public string Content { get; set; }
         public string Author { get; set; }
         public DateTime CreatedAt { get; set; }
         public DateTime UpdatedAt { get; set; }
         public string ClassName { get; set; }
-        public int ObjectId { get; set; }
+        public ulong ObjectId { get; set; }
     }
 
     public partial class JournalContext: DbContext
